test(e2e): add visual snapshots for md-icon-button interaction states
- include snapshots for icon button toggle interaction states on Linux
- add snapshots for icon button toolbar layout on Linux
