+++ tests/e2e/storybook/md-button-family.spec.ts
@@ -1,0 +1,303 @@
+import { expect, test } from '@playwright/test';
+import type { Locator, Page } from '@playwright/test';
+import { openStory } from './storybook.testUtils';
+
+/**
+ * Shared by each button family's real-focus-visible test: waits for the global focus indicator
+ * to reposition, then asserts it matches the focused host's geometry and corner radius, and that
+ * its visible outline extent (outline-width + outline-offset) stays within the viewport.
+ * @param page - The Playwright page used by the behavior test.
+ * @param indicator - The shared `.md-focus-indicator` element.
+ * @param host - The focused component host.
+ */
+const assertFocusIndicatorFollowsHost = async (page: Page, indicator: Locator, host: Locator) => {
+  await page.evaluate(
+    () =>
+      new Promise<void>((resolve) =>
+        requestAnimationFrame(() =>
+          requestAnimationFrame(() => {
+            resolve();
+          }),
+        ),
+      ),
+  );
+
+  const indicatorBox = await indicator.boundingBox();
+  const hostBox = await host.boundingBox();
+  const viewport = page.viewportSize();
+
+  if (!indicatorBox || !hostBox || !viewport) {
+    throw new Error('Missing bounding boxes for focus indicator test.');
+  }
+
+  expect(indicatorBox.width).toBeGreaterThan(0);
+  expect(indicatorBox.height).toBeGreaterThan(0);
+
+  // Geometry follows the rendered host container (allowing for the indicator's outer offset).
+  const TOLERANCE = 1;
+  expect(indicatorBox.x).toBeLessThanOrEqual(hostBox.x + TOLERANCE);
+  expect(indicatorBox.y).toBeLessThanOrEqual(hostBox.y + TOLERANCE);
+  expect(indicatorBox.x + indicatorBox.width).toBeGreaterThanOrEqual(
+    hostBox.x + hostBox.width - TOLERANCE,
+  );
+  expect(indicatorBox.y + indicatorBox.height).toBeGreaterThanOrEqual(
+    hostBox.y + hostBox.height - TOLERANCE,
+  );
+
+  // The indicator's border-radius tracks the focused host's own rendered corner radius.
+  const indicatorRadius = await indicator.evaluate((el) =>
+    parseFloat(getComputedStyle(el).borderRadius),
+  );
+  const hostRadius = await host.evaluate((el) => parseFloat(getComputedStyle(el).borderRadius));
+  expect(indicatorRadius).toBeCloseTo(hostRadius, 0);
+
+  // Clipping must account for the indicator's *visible* outline extent (outline-width +
+  // outline-offset), not just its positioned box, since the outline paints outward from it.
+  const { outlineWidth, outlineOffset } = await indicator.evaluate((el) => {
+    const style = getComputedStyle(el);
+    return {
+      outlineWidth: parseFloat(style.outlineWidth),
+      outlineOffset: parseFloat(style.outlineOffset),
+    };
+  });
+  const visibleExtent = outlineWidth + outlineOffset;
+
+  expect(indicatorBox.x - visibleExtent).toBeGreaterThanOrEqual(0);
+  expect(indicatorBox.y - visibleExtent).toBeGreaterThanOrEqual(0);
+  expect(indicatorBox.x + indicatorBox.width + visibleExtent).toBeLessThanOrEqual(viewport.width);
+  expect(indicatorBox.y + indicatorBox.height + visibleExtent).toBeLessThanOrEqual(viewport.height);
+};
+
+test('MDButton expanded target activates clicks outside the visible button box', async ({
+  page,
+}) => {
+  await openStory(page, 'material-3-components-buttons-mdbutton--expanded-target-hit-area');
+
+  const surface = page.locator('#visual-md-button-target-hit');
+  const button = surface.getByRole('button', { name: 'OK', exact: true });
+  const target = surface.locator('.md-button__target');
+  const count = page.locator('#visual-md-button-target-hit-count');
+  const buttonBox = await button.boundingBox();
+  const targetBox = await target.boundingBox();
+
+  expect(buttonBox).not.toBeNull();
+  expect(targetBox).not.toBeNull();
+
+  if (buttonBox == null || targetBox == null) {
+    throw new Error('Missing MDButton bounding boxes for expanded target hit test.');
+  }
+
+  const clickPoint = {
+    x: buttonBox.x + buttonBox.width / 2,
+    y: buttonBox.y - 2,
+  };
+
+  expect(clickPoint.x).toBeGreaterThan(targetBox.x);
+  expect(clickPoint.x).toBeLessThan(targetBox.x + targetBox.width);
+  expect(clickPoint.y).toBeGreaterThan(targetBox.y);
+  expect(clickPoint.y).toBeLessThan(targetBox.y + targetBox.height);
+  expect(clickPoint.y).toBeLessThan(buttonBox.y);
+
+  await page.mouse.click(clickPoint.x, clickPoint.y);
+
+  await expect(count).toHaveText('1');
+});
+
+test('MDIconButton expanded target activates clicks outside the visible button box', async ({
+  page,
+}) => {
+  await openStory(page, 'material-3-components-buttons-mdiconbutton--expanded-target-hit-area');
+
+  const surface = page.locator('#visual-md-icon-button-target-hit');
+  const button = surface.getByRole('button', { name: 'Expanded target', exact: true });
+  const target = surface.locator('.md-icon-button__target');
+  const count = page.locator('#visual-md-icon-button-target-hit-count');
+  const buttonBox = await button.boundingBox();
+  const targetBox = await target.boundingBox();
+
+  expect(buttonBox).not.toBeNull();
+  expect(targetBox).not.toBeNull();
+
+  if (buttonBox == null || targetBox == null) {
+    throw new Error('Missing MDIconButton bounding boxes for expanded target hit test.');
+  }
+
+  const clickPoint = {
+    x: buttonBox.x - 2,
+    y: buttonBox.y + buttonBox.height / 2,
+  };
+
+  expect(clickPoint.x).toBeGreaterThan(targetBox.x);
+  expect(clickPoint.x).toBeLessThan(targetBox.x + targetBox.width);
+  expect(clickPoint.y).toBeGreaterThan(targetBox.y);
+  expect(clickPoint.y).toBeLessThan(targetBox.y + targetBox.height);
+
+  await page.mouse.click(clickPoint.x, clickPoint.y);
+
+  await expect(count).toHaveText('1');
+});
+
+test('MDIconButton dense toolbar buttons keep click ownership near adjacent boundaries', async ({
+  page,
+}) => {
+  await openStory(page, 'material-3-components-buttons-mdiconbutton--dense-toolbar-interaction');
+
+  const surface = page.locator('#visual-md-icon-button-toolbar-interaction');
+  const addButton = surface.getByRole('button', { name: 'add', exact: true });
+  const filterButton = surface.getByRole('button', { name: 'filter', exact: true });
+  const addCount = page.locator('#toolbar-count-add');
+  const filterCount = page.locator('#toolbar-count-filter');
+  const addBox = await addButton.boundingBox();
+  const filterBox = await filterButton.boundingBox();
+
+  expect(addBox).not.toBeNull();
+  expect(filterBox).not.toBeNull();
+
+  if (addBox == null || filterBox == null) {
+    throw new Error('Missing MDIconButton bounding boxes for dense toolbar edge-click test.');
+  }
+
+  await page.mouse.click(addBox.x + addBox.width / 2, addBox.y + addBox.height / 2);
+  await expect(addCount).toHaveText('1');
+  await expect(filterCount).toHaveText('0');
+
+  await page.mouse.click(filterBox.x + filterBox.width / 2, filterBox.y + filterBox.height / 2);
+  await expect(addCount).toHaveText('1');
+  await expect(filterCount).toHaveText('1');
+
+  await page.mouse.click(addBox.x + addBox.width - 1, addBox.y + addBox.height / 2);
+  await expect(addCount).toHaveText('2');
+  await expect(filterCount).toHaveText('1');
+
+  await page.mouse.click(filterBox.x + 1, filterBox.y + filterBox.height / 2);
+  await expect(addCount).toHaveText('2');
+  await expect(filterCount).toHaveText('2');
+});
+
+test('MDIconButton dense toolbar hover handoff does not leave stale hover state', async ({
+  page,
+}) => {
+  await openStory(page, 'material-3-components-buttons-mdiconbutton--dense-toolbar-interaction');
+
+  const surface = page.locator('#visual-md-icon-button-toolbar-interaction');
+  const addButton = surface.getByRole('button', { name: 'add', exact: true });
+  const filterButton = surface.getByRole('button', { name: 'filter', exact: true });
+  const hovered = page.locator('#toolbar-hovered-button');
+
+  await addButton.hover();
+  await expect(hovered).toHaveText('add');
+  await expect(addButton).toHaveClass(/md-state_hover/);
+
+  await filterButton.hover();
+  await expect(hovered).toHaveText('filter');
+  await expect(filterButton).toHaveClass(/md-state_hover/);
+  await expect(addButton).not.toHaveClass(/md-state_hover/);
+});
+
+test('MDButton focus indicator follows real keyboard focus and is not clipped', async ({
+  page,
+}) => {
+  await openStory(page, 'material-3-components-buttons-mdbutton--focus-indicator-target');
+
+  const host = page.getByRole('button', { name: 'Focus target', exact: true });
+  const indicator = page.locator('.md-focus-indicator');
+
+  // Wait for the story to finish mounting before driving Tab, otherwise the fixed-position
+  // fixture may not yet be an eligible tab stop.
+  await expect(host).toBeVisible();
+
+  // Tab from a page with no focused element: the browser focuses the first focusable element.
+  await page.keyboard.press('Tab');
+  await expect(host).toBeFocused();
+  expect(await host.evaluate((el) => el.matches(':focus-visible'))).toBe(true);
+  await expect(indicator).toHaveCSS('opacity', '1');
+
+  await assertFocusIndicatorFollowsHost(page, indicator, host);
+});
+
+test('MDIconButton focus indicator follows real keyboard focus and is not clipped', async ({
+  page,
+}) => {
+  await openStory(page, 'material-3-components-buttons-mdiconbutton--focus-indicator-target');
+
+  const host = page.getByRole('button', { name: 'Focus target', exact: true });
+  const indicator = page.locator('.md-focus-indicator');
+
+  await expect(host).toBeVisible();
+
+  await page.keyboard.press('Tab');
+  await expect(host).toBeFocused();
+  expect(await host.evaluate((el) => el.matches(':focus-visible'))).toBe(true);
+  await expect(indicator).toHaveCSS('opacity', '1');
+
+  await assertFocusIndicatorFollowsHost(page, indicator, host);
+});
+
+test('MDFab focus indicator follows real keyboard focus and is not clipped', async ({ page }) => {
+  await openStory(page, 'material-3-components-buttons-mdfab--focus-indicator-target');
+
+  const host = page.getByRole('button', { name: 'Focus target', exact: true });
+  const indicator = page.locator('.md-focus-indicator');
+
+  await expect(host).toBeVisible();
+
+  await page.keyboard.press('Tab');
+  await expect(host).toBeFocused();
+  expect(await host.evaluate((el) => el.matches(':focus-visible'))).toBe(true);
+  await expect(indicator).toHaveCSS('opacity', '1');
+
+  await assertFocusIndicatorFollowsHost(page, indicator, host);
+});
+
+test('MDExtendedFab focus indicator follows real keyboard focus and is not clipped', async ({
+  page,
+}) => {
+  await openStory(page, 'material-3-components-buttons-mdextendedfab--focus-indicator-target');
+
+  const host = page.getByRole('button', { name: 'Focus target', exact: true });
+  const indicator = page.locator('.md-focus-indicator');
+
+  await expect(host).toBeVisible();
+
+  await page.keyboard.press('Tab');
+  await expect(host).toBeFocused();
+  expect(await host.evaluate((el) => el.matches(':focus-visible'))).toBe(true);
+  await expect(indicator).toHaveCSS('opacity', '1');
+
+  await assertFocusIndicatorFollowsHost(page, indicator, host);
+});
+
+test('MDIconButton selected pressed shape wins over selected shape under a real pointer press', async ({
+  page,
+}) => {
+  await openStory(page, 'material-3-components-buttons-mdiconbutton--geometry');
+
+  const readRadius = (testId: string) =>
+    page.getByTestId(testId).evaluate((el) => parseFloat(getComputedStyle(el).borderRadius));
+
+  const selectedButton = page.getByTestId('geometry-round-selected');
+  const selectedOnlyRadius = await readRadius('geometry-round-selected');
+  const pressedOnlyRadius = await readRadius('geometry-round-pressed');
+
+  const box = await selectedButton.boundingBox();
+  if (box == null) {
+    throw new Error('Missing MDIconButton bounding box for real-press shape precedence test.');
+  }
+
+  await page.mouse.move(box.x + box.width / 2, box.y + box.height / 2);
+  await page.mouse.down();
+  await expect(selectedButton).toHaveClass(/md-state_pressed/);
+
+  // A real native `:active` press on a selected icon button must still render the plain pressed
+  // shape, matching MDButton's equivalent precedence, not the selected shape. The shape morph is
+  // an animated `border-radius` transition, so poll for the settled value instead of an arbitrary
+  // sleep.
+  await expect
+    .poll(() => readRadius('geometry-round-selected'), { timeout: 2_000 })
+    .toBeCloseTo(pressedOnlyRadius, 5);
+
+  const pressedRadius = await readRadius('geometry-round-selected');
+  await page.mouse.up();
+
+  expect(pressedRadius).not.toBeCloseTo(selectedOnlyRadius, 5);
+});
