+++ tests/e2e/visual/shared-ui/md-button-family.spec.ts-snapshots/md-extended-fab-states-linux.png
