+++ tests/e2e/visual/shared-ui/md-button-family.spec.ts-snapshots/md-icon-button-toolbar-layout-linux.png
